+++ packages/basil/src/BrandColors/index.tsx
@@ -3,10 +3,16 @@ import { css } from '@emotion/core';
 
 interface BrandColorCardProps {
   color: string;
+  name: string;
 }
 
-export const BrandColorCard: React.FC<BrandColorCardProps> = ({ color }) => (
+export const BrandColorCard: React.FC<BrandColorCardProps> = ({
+  name,
+  color,
+}) => (
   <div className="ContentCard" css={css({ backgroundColor: color })}>
-    <span>{color}</span>
+    <span>
+      {name} - {color}
+    </span>
   </div>
 );
