--- packages/basil/src/colors.ts
@@ -1,17 +1,0 @@
-export const colors = {
-  /* BRAND COLORS */
-  BrandGreen: '#1db8a4',
-  BrandYellow: '#ffdb7a',
-  BrandBlue: '#27428c',
-  BrandRed: '#ee534f',
-  BrandPink: '#fdd2c6',
-
-  /* E.L James */
-  GreySadSlate: 'rgba(38, 38, 38, 1)',
-  GreySlate: '#323232',
-  GreyWorstDayOfWinter: '#636363',
-  GreyWinter: '#9b9b9b',
-  GreySpring: '#ececec',
-  GreySummer: '#f5f5f5',
-  GreyBackground: '#fafafa',
-};
