+++ packages/basil/src/index.ts
@@ -1,2 +1,1 @@
-export { colors } from './colors';
 export { Button } from './Button';
