+++ packages/common/src/css/core/colors.css
@@ -24,10 +24,4 @@
   /* other brands */
   --color__twitter: #50abf1;
   --color__snapchat: #fff900;
-
-  --color__f18-blue: #1e1c50;
-  --color__f18-teal: #44c1c3;
-  --color__f18-orange: #e46b3d;
-  --color__f18-purple: #4751a3;
-  --color__f18-pink: #bb1f66;
 }
