+++ packages/common/src/css/sanguine.css
@@ -62,7 +62,6 @@
 @import './components/ImageUpload.css';
 @import './components/Popover.css';
 @import './components/ContentSidebar.css';
-@import './components/FreshersSite.css';
 
 @import '../components/NewsletterSignup/NewsletterSignup.css';
 @import '../components/OrganisationGrid/OrganisationGrid.css';
