+++ packages/falmer/src/components/FalmerHeader/index.tsx
@@ -3,7 +3,7 @@ import { NavLink } from 'react-router-dom';
 import { RootState } from '@ussu/common/src/types/falmer';
 import StudentsUnionLogoNoLogotype from '@ussu/common/src/icons/students-union-logo-no-logotype.svg';
 import { useMappedState } from 'redux-react-hook';
-import { type, Typeface, TypeSize } from '@ussu/common/src/libs/style/type';
+import { type, Typeface, TypeSize } from '@ussu/basil/src/style/type';
 import { css } from '@emotion/core';
 
 const logotypeStyle = css({
