+++ packages/website/src/components/ContentBrowser/ContentBrowserItem.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { ContentBrowserPage } from '../../pages/content/types';
 import { InternalAppLink } from '../InternalAppLink';
-import { COLORS } from '@ussu/common/src/libs/style';
-import { type, Typeface, TypeSize } from '@ussu/common/src/libs/style/type';
+import { COLORS } from '@ussu/basil/src/style';
+import { type, Typeface, TypeSize } from '@ussu/basil/src/style/type';
 import { css } from '@emotion/core';
 
 export const itemActive = css();
