+++ packages/website/src/components/Sectionbar/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { COLORS, MQ, TYPE_SECONDARY } from '@ussu/common/src/libs/style';
+import { COLORS, MQ, TYPE_SECONDARY } from '@ussu/basil/src/style';
 
 export const SectionbarItem: React.FC<
   { active?: boolean; disabled?: boolean } & React.HTMLAttributes<HTMLLIElement>
