+++ packages/website/src/pages/content/pages/FreshersHomepage.tsx
@@ -11,9 +11,9 @@ import UnionLogo from '@ussu/common/src/icons/UnionLogo.svg';
 import { SocialArray } from '../../../components/SocialArray';
 import slugify from '@ussu/common/src/libs/slugify';
 import FreshersWeekArt from '../../../icons/freshersWeek2019ArtEdit.svg';
-import { type, Typeface, TypeSize } from '@ussu/common/src/libs/style/type';
+import { type, Typeface, TypeSize } from '@ussu/basil/src/style/type';
 import StreamField from '../StreamField';
-import { COLORS, MQ } from '@ussu/common/src/libs/style';
+import { COLORS, MQ } from '@ussu/basil/src/style';
 import { useCountdown } from '../../../hooks/useCountdown';
 import { useWindowScroll } from 'react-use';
 import sand from '../../../img/freshers-sand.svg';
