+++ packages/website/src/pages/events/EventDetailPage/EventSidebarSection.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { type, Typeface, TypeSize } from '@ussu/common/src/libs/style/type';
-import { COLORS } from '@ussu/common/src/libs/style';
+import { type, Typeface, TypeSize } from '@ussu/basil/src/style/type';
+import { COLORS } from '@ussu/basil/src/style';
 
 interface EventSidebarSection {
   heading: string;
