+++ packages/website/src/pages/events/EventsApplication/branding/components.tsx
@@ -3,8 +3,8 @@ import { WhatsOnBrandedComponentLocation } from './locations';
 import React from 'react';
 import EVENT_LISTINGS_BUNDLE from '../../EventsCalender/EventListingsBundle.graphql';
 import { useQuery } from '@apollo/react-hooks';
-import { type, TypeSize } from '@ussu/common/src/libs/style/type';
-import { Layers, MQ } from '@ussu/common/src/libs/style';
+import { type, TypeSize } from '@ussu/basil/src/style/type';
+import { Layers, MQ } from '@ussu/basil/src/style';
 import FauxRouterLink from '../../../../components/FauxRouterLink';
 import { BuyButton } from '../../EventDetailPage/BuyButton';
